Pull validateDir out into a new App.Util module
