+++ src/App/Fossa/Main.hs
@@ -9,11 +9,12 @@ import App.Fossa.Analyze (ScanDestination (..), analyzeMain)
 import App.Fossa.FossaAPIV1 (ProjectMetadata (..))
 import App.Fossa.Report (ReportType (..), reportMain)
 import App.Fossa.Test (TestOutputType (..), testMain)
+import App.Util (validateDir)
 import qualified Data.Text as T
 import OptionExtensions
 import Effect.Logger
 import Options.Applicative
-import Path.IO (doesDirExist, resolveDir', setCurrentDir)
+import Path.IO (setCurrentDir)
 import Prologue
 import System.Environment (lookupEnv)
 import System.Exit (die)
@@ -46,7 +47,6 @@ appMain = do
       changeDir reportBaseDir
       key <- requireKey maybeApiKey
       reportMain optBaseUrl key logSeverity reportTimeout reportType optProjectName optProjectRevision
-      
 
 requireKey :: Maybe Text -> IO Text
 requireKey (Just key) = pure key
@@ -59,16 +59,6 @@ changeDir path = validateDir path >>= setCurrentDir
 checkAPIKey :: Maybe Text -> IO (Maybe Text)
 checkAPIKey Nothing = fmap T.pack <$> lookupEnv "FOSSA_API_KEY"
 checkAPIKey key = return key
-
--- | Validate that a filepath points to a directory and the directory exists
-validateDir :: FilePath -> IO (Path Abs Dir)
-validateDir dir = do
-  absolute <- resolveDir' dir
-  exists <- doesDirExist absolute
-
-  unless exists (die $ "ERROR: Directory " <> show absolute <> " does not exist")
-
-  pure absolute
 
 baseDirArg :: Parser String
 baseDirArg = argument str (metavar "DIR" <> help "Set the base directory for scanning (default: current directory)" <> value ".")
@@ -140,7 +130,7 @@ reportOpts =
     <*> baseDirArg
 
 reportCmd :: Parser ReportType
-reportCmd = 
+reportCmd =
   hsubparser $
     command "attribution" (info (pure AttributionReport) $ progDesc "Generate attribution report" )
